+++ src/fides/api/ctl/migrations/versions/2fb48b0e268b_update_ctl_datasets_fidesctl_meta.py
@@ -1,7 +1,7 @@
 """update ctl_datasets fidesctl_meta
 
 Revision ID: 2fb48b0e268b
-Revises: a08024ba7e2b
+Revises: 5b4b9c2d1c93
 Create Date: 2022-12-08 17:49:14.317905
 
 """
@@ -11,7 +11,7 @@ from sqlalchemy.dialects import postgresql
 
 # revision identifiers, used by Alembic.
 revision = "2fb48b0e268b"
-down_revision = "a08024ba7e2b"
+down_revision = "5b4b9c2d1c93"
 branch_labels = None
 depends_on = None