+++ src/fides/api/ctl/routes/crud.py
@@ -67,7 +67,7 @@ async def validate_data_categories(
 
 
 # CRUD Endpoints
-routers = []
+routers: List[APIRouter] = []
 for model_type, fides_model in model_map.items():
     # Programmatically define routers for each resource type
     router = APIRouter(